requirements.txt
psycopg2-binary>=2.9.0
google-genai>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
chromadb>=0.5.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
beautifulsoup4>=4.12.0
pytest>=8.0.0
pytest-asyncio>=0.21.0
httpx>=0.27.0
# Optional: PDF ingestion
# pypdf>=4.0.0
